+++ src/main.c
@@ -13,12 +13,12 @@
 int main ( void )
 {
     /* Initialize all modules */
-    SYS_Initialize ( NULL );
+    SYS_Initialize ( NULL );      
 
     while ( true )
     {
         /* Maintain state machines of all polled MPLAB Harmony modules. */
-        SYS_Tasks ( );
+        SYS_Tasks ( );        
     }
 
     /* Execution should not come here during normal operation */