+++ bootloader/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Node90 Bootloader Firmware
 
+# 1.1.2
+- add LED_STAT_Toggle() during firmware update
+- change WDTPS PS8192 8 seconds timeout
+
 # 1.1.1
 - fix binary size +1